Split out signal tests
